src/main/java/com/life/full/danger/sell/product/service/impl/ProductServiceImpl.java
package com.life.full.danger.sell.product.service.impl;

import com.life.full.danger.sell.product.ProductInfoDao;
import com.life.full.danger.sell.product.enums.ProductStatusEnum;
import com.life.full.danger.sell.product.model.dbobject.ProductInfo;
import com.life.full.danger.sell.product.service.ProductService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * @author LiHongHui
 * @date 2018/9/14 10:24
 * @description
 */
@Service
public class ProductServiceImpl implements ProductService {
    @Autowired
    private ProductInfoDao productInfoDao;

    @Override
    public ProductInfo findOne(String productId) {
        return productInfoDao.findById(productId).get();
    }

    /**
     * 查询所有在使用的商品
     *
     * @return
     */
    @Override
    public List<ProductInfo> findUpAll() {
        return productInfoDao.findByProductStatus(ProductStatusEnum.UP.getCode());
    }

    @Override
    public Page<ProductInfo> findAll(Pageable pageable) {
        return productInfoDao.findAll(pageable);
    }

    @Override
    public ProductInfo save(ProductInfo productInfo) {
        return productInfoDao.save(productInfo);
    }
}
